Remove System.Runtime.Loader ref and bump DI version
We used System.Runtime.Loader because we use `AssemblyLoadContext`, and you need that package for that to be available on .NET Standard. Since we're dropping .NET Standard support, we're losing that ref, and also updating to a v8 Microsoft.Extensions.DependencyInjection.Abstractions so that we're all aligned with .NET >= 8.0. (Not strictly necessary, but it tends to confuse people when you use older-looking versions.)